signed <- unsigned

+++ Marlin/src/lcd/e_paper/libs/U8g2_for_Adafruit_GFX/src/U8g2_for_Adafruit_GFX.h
@@ -36,7 +36,6 @@
 #ifndef U8g2_for_Adafruit_GFX_h
 #define U8g2_for_Adafruit_GFX_h
 
-//#include <Adafruit_GFX.h>
 #include "../../Adafruit_GFX/Adafruit_GFX.h"
 #include "u8g2_fonts.h"
 
@@ -58,16 +57,16 @@ struct _u8g2_font_info_t
   uint8_t bits_per_delta_x;
 
   /* offset 9 */
-  uint8_t max_char_width;
-  uint8_t max_char_height; /* overall height, NOT ascent. Instead ascent = max_char_height + y_offset */
-  uint16_t x_offset;
-  uint16_t y_offset;
+  int8_t max_char_width;
+  int8_t max_char_height; /* overall height, NOT ascent. Instead ascent = max_char_height + y_offset */
+  int8_t x_offset;
+  int8_t y_offset;
 
   /* offset 13 */
-  uint8_t  ascent_A;
-  uint8_t  descent_g;  /* usually a negative value */
-  uint8_t  ascent_para;
-  uint8_t  descent_para;
+  int8_t  ascent_A;
+  int8_t  descent_g;  /* usually a negative value */
+  int8_t  ascent_para;
+  int8_t  descent_para;
 
   /* offset 17 */
   uint16_t start_pos_upper_A;
@@ -82,15 +81,15 @@ struct _u8g2_font_decode_t
 {
   const uint8_t *decode_ptr;      /* pointer to the compressed data */
 
-  uint16_t target_x;
-  uint16_t target_y;
+  int16_t target_x;
+  int16_t target_y;
   uint16_t fg_color;
   uint16_t bg_color;
 
-  uint16_t x;           /* local coordinates, (0,0) is upper left */
-  uint16_t y;
-  uint8_t glyph_width;
-  uint8_t glyph_height;
+  int8_t x;           /* local coordinates, (0,0) is upper left */
+  int8_t y;
+  int8_t glyph_width;
+  int8_t glyph_height;
 
   uint8_t decode_bit_pos;     /* bitpos inside a byte of the compressed data */
   uint8_t is_transparent;
@@ -114,17 +113,17 @@ struct _u8g2_font_t
   //int8_t font_ref_ascent;
   //int8_t font_ref_descent;
 
-  uint8_t glyph_x_offset;    /* set by u8g2_GetGlyphWidth as a side effect */
+  int8_t glyph_x_offset;    /* set by u8g2_GetGlyphWidth as a side effect */
 };
 typedef struct _u8g2_font_t u8g2_font_t;
 
 
 uint8_t u8g2_IsGlyph(u8g2_font_t *u8g2, uint16_t requested_encoding);
-uint8_t u8g2_GetGlyphWidth(u8g2_font_t *u8g2, uint16_t requested_encoding);
+int8_t u8g2_GetGlyphWidth(u8g2_font_t *u8g2, uint16_t requested_encoding);
 void u8g2_SetFontMode(u8g2_font_t *u8g2, uint8_t is_transparent);
 void u8g2_SetFontDirection(u8g2_font_t *u8g2, uint8_t dir);
-uint16_t u8g2_DrawGlyph(u8g2_font_t *u8g2, uint16_t x, uint16_t y, uint16_t encoding);
-uint16_t u8g2_DrawStr(u8g2_font_t *u8g2, uint16_t x, uint16_t y, const char *s);
+int16_t u8g2_DrawGlyph(u8g2_font_t *u8g2, int16_t x, int16_t y, uint16_t encoding);
+int16_t u8g2_DrawStr(u8g2_font_t *u8g2, int16_t x, int16_t y, const char *s);
 void u8g2_SetFont(u8g2_font_t *u8g2, const uint8_t  *font);
 void u8g2_SetForegroundColor(u8g2_font_t *u8g2, uint16_t fg);
 void u8g2_SetBackgroundColor(u8g2_font_t *u8g2, uint16_t bg);
@@ -133,14 +132,14 @@ void u8g2_SetBackgroundColor(u8g2_font_t *u8g2, uint16_t bg);
 class U8G2_FOR_ADAFRUIT_GFX : public Print {
   public:
     u8g2_font_t u8g2;
-    uint16_t tx, ty;       // current position for the print command
+    int16_t tx, ty;       // current position for the print command
     uint16_t encoding;    // the unicode, detected by the utf-8 decoder
     uint8_t utf8_state;   // current state of the utf-8 decoder, contains the remaining bytes for a detected unicode glyph
 
     void home() { tx = 0; ty = 0; utf8_state = 0; }
     void setCursor(uint16_t x, uint16_t y) { tx = x; ty = y; utf8_state = 0; }
-    uint16_t getCursorX() { return tx; }
-    uint16_t getCursorY() { return ty; }
+    int16_t getCursorX() { return tx; }
+    int16_t getCursorY() { return ty; }
 
     U8G2_FOR_ADAFRUIT_GFX() {u8g2.font = NULL; u8g2.font_decode.fg_color = 1; u8g2.font_decode.is_transparent = 1; u8g2.font_decode.dir = 0; home(); }
     void begin(Adafruit_GFX &gfx) { u8g2.gfx = &gfx; }
@@ -154,23 +153,23 @@ class U8G2_FOR_ADAFRUIT_GFX : public Print {
       { u8g2_SetForegroundColor(&u8g2, fg); }
     void setBackgroundColor(uint16_t bg)       // only used for setFontMode(0)
       { u8g2_SetBackgroundColor(&u8g2, bg); }
-    uint8_t getFontAscent()
+    int8_t getFontAscent()
       { return u8g2.font_info.ascent_A; }
-    uint8_t getFontDescent()
+    int8_t getFontDescent()
       { return u8g2.font_info.descent_g; }
-    uint16_t drawGlyph(uint16_t x, uint16_t y, uint16_t e)
+    int16_t drawGlyph(int16_t x, int16_t y, uint16_t e)
       { return u8g2_DrawGlyph(&u8g2, x, y, e); }  // draw a signle char (e == Unicode)
-    uint16_t drawStr(uint16_t x, uint16_t y, const char *s)
+    int16_t drawStr(int16_t x, int16_t y, const char *s)
       { return u8g2_DrawStr(&u8g2, x, y, s); }
 
-    uint16_t drawUTF8(uint16_t x, uint16_t y, const char *str);
-    uint16_t getUTF8Width(const char *str);
+    int16_t drawUTF8(int16_t x, int16_t y, const char *str);
+    int16_t getUTF8Width(const char *str);
 
     uint16_t utf8_next(uint8_t b);
 
     size_t write(uint8_t v) {
       uint16_t e = utf8_next(v);
-      uint16_t delta;
+      int16_t delta;
 
       if ( e == '\n' ) {
         tx = 0;
@@ -1622,7 +1621,4 @@ extern const uint8_t u8g2_font_px437wyse700b_mn[] U8G2_FONT_SECTION("u8g2_font_p
 
 /* end font list */
 
-
 #endif /* U8g2_for_Adafruit_GFX_h */
-
-
